Restructure server folders
