+++ server/package.json
@@ -4,14 +4,17 @@
   "description": "Backend for rattle-battle",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
-    "lint": "npx eslint index.js",
+    "start": "node src/index.js",
+    "start:dev": "nodemon src",
+    "lint": "npx eslint src/index.js",
     "test": "jest"
   },
   "author": "Jithin Tom <jithintom1@gmail.com>,",
   "license": "MIT",
   "private": false,
   "dependencies": {
+    "body-parser": "^1.18.3",
+    "dotenv": "^6.2.0",
     "express": "^4.16.4"
   },
   "devDependencies": {
