+++ server/src/routes/routes.js
@@ -1,0 +1,10 @@
+
+const express = require('express');
+
+const router = express.Router();
+
+router.get('/ping', (req, res) => {
+  res.json({ status: 'working' });
+});
+
+module.exports = router;
